requirements.txt
playwright
google-generativeai
google-genai
python-dotenv
beautifulsoup4
